int --> long long int

+++ functions.c
@@ -1,11 +1,11 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-bool prime1(int number){
-    int i = 2;
+bool prime1(long long int number){
+    long long int  i = 2LL;
     bool isPrime = true && number != 1;
     while(i < number && isPrime){
-        if(number % i == 0){
+        if(number % i == 0LL){
             isPrime = false;
         }
         else i++;
@@ -14,7 +14,7 @@ bool prime1(int number){
 }
 
 bool prime2(int number){
-    int i = 2;
+    long long int i = 2LL;
     bool isPrime = true && number != 1;
     while(i <= (number / 2) && isPrime){
         if(number % i == 0){
@@ -26,7 +26,7 @@ bool prime2(int number){
 }
 
 bool prime3(int number){
-    int i;
+    long long int i;
     bool isPrime = true;
     if((number != 2 && number % 2 == 0) ||  number == 1)
         isPrime = false;
@@ -43,7 +43,7 @@ bool prime3(int number){
 }
 
 bool prime4(int number){
-    int i;
+    long long int i;
     bool isPrime = true;
     if((number != 2 && number % 2 == 0) ||  number == 1)
         isPrime = false;
@@ -61,7 +61,7 @@ bool prime4(int number){
 
 bool prime5(int number){
     bool isPrime = true && number != 1;
-    int i = 2;
+    long long int i = 2LL;
     while(i <= sqrt(number) && isPrime){
         if(number % i == 0){
             isPrime = false;
@@ -72,12 +72,12 @@ bool prime5(int number){
 }
 
 bool prime6(int number){
-    int i;
+    long long int i;
     bool isPrime = true;
     if((number != 2 && number % 2 == 0) ||  number == 1)
         isPrime = false;
     else if(number != 2){
-        i = 3;
+        i = 3LL;
         while(i <= sqrt(number) && isPrime){
             if(number % i == 0){
                 isPrime = false;
@@ -87,3 +87,5 @@ bool prime6(int number){
     }
     return isPrime;
 }
+
+
